Test : /palce/delete api test case add
삭제하려는 장소가 존재하지 않을 때 실패 케이스 추가

+++ server/src/main/java/onde/there/place/service/PlaceService.java
@@ -22,10 +22,15 @@ public class PlaceService {
 
 	@Transactional
 	public boolean delete(Long placeId) {
+		boolean exists = placeRepository.existsById(placeId);
+		if (!exists) {
+			throw new PlaceException(ErrorCode.NOT_FOUND_PLACE);
+		}
+
 		placeRepository.deleteById(placeId);
 
 		//TODO: image 제거 로직 -> 이미지 추가 삭제 부분 머지후 구현 예정
 
-		return !placeRepository.existsById(placeId);
+		return true;
 	}
 }
